Added support for local logging
The local logging driver has native support since Docker 1.19.3.
This new logging driver has proven to be stable and was implemented
to be more efficient both at disk space consumption and when
parsing and reading entire log files.

Since this is now the logging driver that Docker team recommends
for most use-cases, we should support it in Swarmpit too.

Turn out, since `docker logs` has native support to it, it was just
a matter of enabling the 'local' word in some places.